[level 1] Title: x만큼 간격이 있는 n개의 숫자, Time: 0.04 ms, Memory: 79.5 MB -BaekjoonHub

+++ 프로그래머스/1/12954. x만큼 간격이 있는 n개의 숫자/x만큼 간격이 있는 n개의 숫자.java
@@ -1,9 +1,9 @@
 class Solution {
-    public long[] solution(long x, int n) {
+    public long[] solution(int x, int n) {
         long[] answer = new long[n];
         
         int i = 0;
-        long plus = x;
+        int plus = x;
         while (i < n) {
             answer[i] = x;
             x += plus;
